cros-kernel2.eclass: Parallelize arm64 kernel compression
During an emerge of the kernel on arm64 we spend a chunk of time
waiting for the kernel to compress.  On a kasan-enabled build (on
4.19) I measured this at about 2.7 seconds worth of time.  On a
non-kasan enabled build (also 4.19) this seems to be about 1.5
seconds.  The difference is, presumably, the overall size of the
kernel.

It turns out that there are some other tasks (modules_install, which
we call twice, and getting the kernel release) that can happen in
parallel.  At the moment those tasks are pretty slow (~3.3 seconds
each when using clang).  Let's let those tasks happen in parallel with
compressing.  Future changes may reduce the overhead of the above, but
it's unlikely we'll manage to reduce it below 2-3 seconds so it's
still worth doing things in parallel.

Since I'm touching the fit generation / compression code, let's also
adjust it to take some of the coding style suggestions made in other
cros-kernel2 CLs.

BUG=chromium:767073
TEST=Build/install a kernel on cheza
TEST=Build/install a kernel on veyron_jerry
TEST=Build w/ -fit_compression_kernel_lz4 fit_compression_kernel_lzma
TEST=Build w/ -fit_compression_kernel_lz4 and boot on cheza
TEST=Build a kernel on samus (non-arm64)

Change-Id: Ida41b3f4653d12797645927720058913c5e88b21
Reviewed-on: https://chromium-review.googlesource.com/1518796
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>