cros-rust: export RUST_BACKTRACE=1
This is something we want for all packages so set it in the eclass
rather than requiring each package to set it individually.

However, there is a leak in the library that rust uses to get backtraces
so don't set it when building with asan.

BUG=chromium:940546
TEST=`FEATURES=test USE=asan emerge-$BOARD crosvm` works and doesn't
      report any leaks

Change-Id: Id4dc7f77684742699eb7b3831515ae7bffe7a791
Reviewed-on: https://chromium-review.googlesource.com/1518669
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>