journald: disable MaxFileSec
By default, MaxFileSec is set to 1 month, and any journal file older
than a month is rotated. This default is in place to avoid losing too
much data at once when an old file is deleted. Since we have a fixed
number of journal files, we don't need this. Having it on decrease
storage density since journal files that are rotated before they're
full have a lot of wasted space in them.

This also works around a systemd bug seen on jestream where journald
crashes due to a SIGBUS in the journal-offline thread. This is necessary
until the root cause of the SIGBUS crash is fixed in upstream and we
upgrade to a version of systemd that has it.

BUG=b:125762558
TEST=Move clock forward by 2 months and ensure the journal file isn't
rotated.
TEST=Boot jestream 10 times in a row without it hanging.

Change-Id: I0ed8ef2251687c9739caae2a4ae60825c64160a5
Reviewed-on: https://chromium-review.googlesource.com/1520086
Commit-Ready: Christopher Morin <cmtm@google.com>
Tested-by: Christopher Morin <cmtm@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>