cros-kernel2.eclass: Fix stripping to be faster and do splitdebug
Back in CL:724168 we fixed symbolization of stack crawls but with two
downsides:
- We now stash entire unstripped modules in the debug directory
  instead of using splitdebug.
- We now have to call into the kernel's "modules_install" twice, which
  is slow (over 3 seconds on clang builds)

Let's fix this by implementing a custom STRIP command.  The custom
command will follow the standard splitdebug recipe and stash away the
.debug file in the right place.  Then it will do the normal kernel
--strip-debug.

With the new way of doing things, we fix all the downsides of the old
method.  That means we get back to a faster build and we also get back
to using splitdebug.

NOTE: it'd be nice to be able to just export PORTAGE_STRIP_FLAGS to
change portage's behavior instead of doing all this hacking around.
However, for whatever reason, when estrip processes "SB relocatable"
files it ignores "PORTAGE_STRIP_FLAGS" and uses "SAFE_STRIP_FLAGS":

https://chromium.googlesource.com/chromiumos/third_party/portage_tool/
  +/refs/heads/chromeos-2.3.49/bin/estrip#438

...and the kernel modules fall into that.  If someone can ever figure
out how to make that work properly (which presumably means
understanding why portage currently behaves the way it does) we can
simplify all this and remove our custom stripping script.

BUG=chromium:767073, chromium:221261
TEST=Hack this to "--strip-unneeded" compare with portage results

Change-Id: Icdf0b7fedb705c3dc502e7ceb34d1fea6d0fd0ef
Reviewed-on: https://chromium-review.googlesource.com/1521528
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>