cros-kernel2.eclass: Fast path getting the kernel release
Calling into the kernel Makefile is super slow and generally we should
try to avoid it whenever we can.  In Chrome OS we're extra slow
because of all the overhead of invoking our compiler, but even if that
gets fixed there will still be a non-negligible overhead.

Let's optimize the "install" case to not call into the kernel Makefile
just to get the release version of the kernel.  This version can be
gleaned from the name of the files that the kernel installs and doing
a bit of shell globbing and string manipulation is much faster.

If somehow our globbing / string manipulation fail (because the kernel
installed two files that look like /boot/vmlinuz-*???) we'll fall back
to the slow way of asking the kernel.  This is more paranoia than
anything I'd expect to happen and perhaps should just be deleted, but
it also doesn't hurt.

Comparing an incremental build on cheza, before:
  0m31.377s, 0m30.883s, 0m31.082s
After:
  0m27.293s, 0m27.275s, 0m27.717s

...so this saves over 3.5 seconds on an incremental build.

BUG=chromium:767073
TEST=Kernel build still works.

Change-Id: Id89f78845c053f50e96f06f20325940d64d4788d
Reviewed-on: https://chromium-review.googlesource.com/1522086
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>