gtest gmock: Transition away from separate packages.
This upgrades gtest to 1.8.1 with the 1.8.0-r2 version as an
intermediate step towards migrating to the 1.8.1 gtest version on
portage stable which integrates gmock into the gtest package.
Consequently, the gmock-2.0.0.0 package is a dummy package that meets
the requirements of packages depending on gmock until they are modified
to depend only on gtest.

BUG=chromium:940320
TEST=gtest and gmock emerge, unit test CQ stage passes locally.

Cq-Depend: chromium:1685899, chromium:1688282
Change-Id: Iedc0dc82addb933ad22ec712b3aa76f59e0ddeee
Reviewed-on: https://chromium-review.googlesource.com/1529259
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Ready: Allen Webb <allenwebb@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>