libxkbcommon: upgraded package to 0.8.2
Upgraded x11-libs/libxkbcommon to version 0.8.2 on amd64 and arm.

Note that we patch the package to provide 0.4.3 symbols version for a
subset of symbols used by Chrome so that we can load it on systems with
updated libxkbcommon. We'll drop them once we get Chrome binaries built
against 0.8.2.

BUG=b:128975017
TEST=Build and boot eve-kvm, verify Chrome comes up

Change-Id: I54d00d88d52e45481514bcac293911afaa3b4760
Reviewed-on: https://chromium-review.googlesource.com/1540536
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Reviewed-by: Kevin Schoedel <kpschoedel@chromium.org>