ureadahead: Upgrade to 0.100.0-13
We've been carrying a patch in our kernel sources to revive the
'tracing_enabled' file in /sys/kernel/debug/tracing because ureadahead
uses it. This problem was fixed in ureadahead version 0.100-0.13, but we
never upgraded to it (see this bug[1]). Let's pull in the diff from the
previous version we synced to, 0.100-0.11, and apply it as a patch so
that we can drop the kernel patch we've been carrying for many years.

BUG=chromium:248268
TEST=Boot up device, see /var/lib/ureadahead/pack is still created

[1] https://bugs.launchpad.net/ubuntu/+source/ureadahead/+bug/1085766
Change-Id: Ia54307895f03bd58425521fd3d7a29851554c5cb
Reviewed-on: https://chromium-review.googlesource.com/1542127
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>