chromeos-chrome: bump size checks; disable for external builds
Non-chrome_internal "release" builds of Chromium (e.g. AFDO, ThinLTO,
CFI, etc. all enabled) are hitting this check, since they clock in at
206MB. The use of chrome_internal drops this to well below this 200MB
threshold.

The biggest contributor to this difference appears to be the addition of
unwinding info. Unwinding info is left in if we're not doing a build of
Chrome that's both official and branded.

Since we're creeping toward 200MB on chrome_internal builds, this also
bumps the suspicious Chrome size up to 250MB.

BUG=None
TEST=emerge-amd64-generic chromeos-chrome

Change-Id: If594cb7352cebaddb9dc3c1884fd7118d3ca1ff5