llvm: Change llvm pgo generate USE flags
From pgo_generate to llvm_pgo_generate.
Because we want to pass this flag through append_useflags in a builder
and it is global. pgo_generate has already been defined in other projects
and will be also turned which we don't want to see.

BUG=None
TEST=None

Change-Id: Ia3ea186dbfd3555cb811700800131b64de309877
Reviewed-on: https://chromium-review.googlesource.com/1519111
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: Zhizhou Yang <zhizhouy@google.com>