Add grpcio Python modules to SDK
We are updating the FAFT RPC server to use gRPC -- see go/faft-grpc.
This change adds the necessary SDK package, dev-python/grpcio.

The DUT will also need to access the grpcio module,
but because the Gentoo package does not work with ARM boards,
we will handle that separately.

We also take this opportunity to upgrade the EAPI of autotest-deps
from 4 to 5, as required by the presubmit queue.

BUG=None
TEST=...
1. Uninstall packages from your SDK and board to ensure a clean test:
`sudo emerge -C dev-python/grpcio-tools-1.19.0
dev-python/grpcio-1.19.0-r1`
2. Verify that your SDK and board do not recognize the modules:
2a. `python -m grpc` should yield "No module named grpc".
2b. Likewise for `python -m grpc_tools`.
3. Run `trunk/src/scripts/build_packages --board=$BOARD`.
4. Verify that your SDK and board now recognize the modules:
5a. `python -m grpc` should yield "No module named grpc.__main__; 'grpc'
is a package and cannot be directly executed".
5b. Likewise for `python -m grpc_tools`.

Change-Id: I7bd163a05458b3d8bbe9d42a8165bbdd8df740b1
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1553972
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>