chromeos-chrome: forcibly deprecate gold
We know of no users of `USE=-lld gold`, but it's likely best to break
them loudly rather than silently, since lld and gold are very compatible
in very many ways. The intent is for this change to cook for a few days
before we remove all of the gold bits in this ebuild Once and For All.

BUG=None
TEST=CQ

Change-Id: I7676e8fa14132d9eefc1d9a3cae1b85396fde466
Reviewed-on: https://chromium-review.googlesource.com/1558655
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>