libqrtr: Support multilib compilation
This library's makefile uses $libdir to determine where to put the
library it builds, but defaults to $prefix/lib if not specified. Portage
wants to put 64-bit libraries in lib64, and it will fail to build if
this isn't the case, so always specify $libdir to be what portage wants
it to be. This fixes the following emerge error for this package:

 Files matching a file type that is not allowed:
    usr/lib/libqrtr.so.1.0

BUG=chromium:950966
TEST=emerge libqrtr on 64-bit userland

Change-Id: I809eada946427e18f481328ca30c37c09b00d5ef
Reviewed-on: https://chromium-review.googlesource.com/1559691
Commit-Ready: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>