net-fs/sshfs: Use a dummy mountpoint mode when using a passed /dev/fuse fd
sshfs verifies that the file type of the mountpoint and the remote file
match (i.e. both are directories). When using a passed /dev/fuse fd, the
mountpoint is /dev/fd/%d, where %d is the /dev/fuse file descriptor.
This file isn't a directory and hence won't match the type of the remote
file.

BUG=chromium:866377
TEST=platform_CrosDisks* autotests
TEST=Enable unprivilaged sshfs in cros-disks and access Linux files in
     the Files app

Change-Id: I5b9cf72a3e79fa15b8cedc6258fa2014a74392ea
Reviewed-on: https://chromium-review.googlesource.com/1588604
Commit-Ready: Anand Mistry <amistry@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>