media-libs/arc-mesa: fix Vulkan test on N
This fixes dEQP-VK.synchronization.events on Nami.  Note
this test was removed from dEQP a while ago, and this
revert is needed to satisfy N CTS.

However, it's interesting to note the test still passes on
a v4.13 Intel setup, while it fails on Nami/Fizz (v4.4 devices).
Could be a real bug, but it's pretty easy to revert either
way.

Also separate non-upstreamable, N specific fixes since we're
eventually going to split the ebuilds.

BUG=b:130195490
TEST=./deqp-vk --deqp-case=dEQP-VK.synchronization.events passes on Nami

Change-Id: Iec503af14c768b527778fed5195ba0fbf3321632