media-libs/mesa: Don't use BGRA internal format on vm build.
Since we only enable virgl driver for vm builds, we just use it as
a flag to indicate if it's for a vm build. Currently we also build
intel driver on tatl board which is used by crostini. We can't force
BGRA internal format on desktop GL since it's a invalid internal
format on desktop GL.

BUG=None
TEST=None

Change-Id: I225d82f38d16e51d4162f70f2de56d657f42b869
Reviewed-on: https://chromium-review.googlesource.com/1568405
Commit-Ready: Lepton Wu <lepton@chromium.org>
Tested-by: Lepton Wu <lepton@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>