dev-rust/pkg-config: patch SYSROOT detection
The Rust pkg-config wrapper (pkg-config-rs) uses the path to a library
to determine whether it is a "system" library.  It bases this
determination on whether the path begins with "/usr", which is not
correct for CrOS cross builds where the sysroot is located in
/build/${BOARD}.

Add a patch to allow overriding the sysroot using the SYSROOT
environment variable, which is set in the CrOS build system, or the
PKG_CONFIG_SYSROOT_DIR variable, which isn't (except inside the
pkg-config-$BOARD helper itself, but we can't use that inside the Rust
wrapper that calls it).

BUG=None
TEST=USE='crosvm-tpm' emerge-nami crosvm

Change-Id: Ibca7f565aab648a215450b10d0d68d2b55ac6b86
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1577752
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>