cros-rust: run default src_prepare to apply patches
The cros-rust eclass overrides src_prepare, but its implementation does
not apply patches from the PATCHES global in the same way that the
default implementation does. Call the default implementation so that
patch application in cros-rust ebuilds works by default. (The call to
eapply_user is also part of the default src_prepare, so drop it.)

BUG=None
TEST=emerge-nami crosvm; ebuild prepare an ebuild with PATCHES

Change-Id: Ia95c5158157f4b10259219d6832dc03ad403e0e8
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1577751
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>