chromeos-chrome: Add USE flag to generate orderfile using new pass manager.
This patch adds a new USE flag "orderfile_generate" in chromeos-chrome ebuild
that controls to generate an orderfile of symbols ordered by Call-Chain
Clustering (C3). This patch is intended to be submitted after crrev.com/c/1650618,
as that patch is making changes in Chrome GN files. This patch is to turn on
the GN arg and do installation.

BUG=chromium:950627
TEST=Tested with orderfile-generation tryjob
https://ci.chromium.org/p/chromeos/builders/general/Try/b8911280322967308192

Change-Id: Iec70d4dc8bc677bf729d8d7f7241ea4532bf5b4a