alsa-lib: Delete the VEYRON-I2S ucm config
In CL:1579953 we want to install UCM configs with the name matching
what's in the upstream kernel.  ...but we can't because it conflicts
with what's coming from alsa-lib.

We can't just use what's in alsa-lib because we have some slightly
different configs for the various veyron models.

Until someone comes up with a better solution, we'll just delete the
files from alsa-lib.

BUG=chromium:941638
TEST=Play the youtube video It's Raining Tacos

Change-Id: I73ccaf624f6fe5034331f158faad93426630c2fd
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1579828
Reviewed-by: Dylan Reid <dgreid@chromium.org>