coreboot: Set sarien/arcada to not touch descriptor regions
In order to ensure the EC region remains unwritable from the host set the
coreboot option to ensure that it does not run ifdtool to fully unlock
all the regions and instead to use the settings that come from the FIT
tool and the initial descriptor region.

This will force all EC updates to go through software sync.

BUG=b:130706387
TEST=build coreboot and ensure EC region is not writable from the host

Change-Id: If7940e9ffae3cb8174d64822ce5a815b4741be88
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1583252
Reviewed-by: Furquan Shaikh <furquan@chromium.org>