chromeos-base/tpm_manager: Pull trousers if !tpm2
Currently it's possible to emerge this ebuild with neither tpm nor tpm2
set, which ends up with a failure due to missing dependencies, because
they are gated on one of the 2 flags. Since the package always includes
TPM supports if TPM2 is not enabled, gate the TPM dependencies on !tpm2.

BUG=none
TEST=USE="-tpm2 mocktpm" ./build_packages --board=scarlet builds fine

Reported-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>

Change-Id: I4468e4d5d2da1bbe341b37c37624a1e43175ad70
Reviewed-on: https://chromium-review.googlesource.com/1583263
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>