Add dev-python/grpcio Gentoo package to autotest
This package will be used for go/faft-grpc.

The upstream dev-python/grpcio package assumes that SYSROOT='/',
which is not always true on our boards.
For example, whirlwind has a different SYSROOT.
Hence, this commit includes a patch to the grpcio package
to read SYSROOT from the environment variables, defaulting to '/'.

BUG=chromium:963160
TEST=ebuild-whirlwind chromeos-base/autotest/autotest-9999.ebuild \
clean install

Change-Id: I1c21cdd6e75604bf209b8a3e91f2527473eddcc7
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1601057
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>