cros-kernel2.eclass: Add support for device tree compression
This patch adds support for the new device tree compression feature for
which firmware support was added to depthcharge in CL:1553458.
Compressed device trees shrink to less than a third of the size and thus
reduce the overall size of the FIT image, increasing boot speed.

This feature is supposed to be the new default for all future Chrome OS
devices, so this patch enables it by default to make sure we don't have
to carry a special USE flag forever and risk forgetting it every time we
add a new board. However, since the feature requires firmware support,
it needs to be explicitly disabled on all previously shipped boards (see
dependent CLs). The only board families that will have it enabled to
start with are Cheza, Kukui and Mistral.

BUG=None
TEST=Built and booted Kevin, confirmed that FDTs were still
uncompressed. Built and booted Kevin with explicit USE=dt_compression,
confirmed that FDTs were compressed and that ToT depthcharge could boot
them correctly.

Cq-Depend: chromium:1621519,chrome-internal:1323952,chrome-internal:1324624,chrome-internal:1323953,chrome-internal:1324499,chrome-internal:1320522
Change-Id: I7651b9b87d91e14e5baa8872f7e6cc444c186bf9
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1621487
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>