dev-lang/rust: update upgrade instructions
Change the instructions to keep the latest two versions (to avoid
problems in the time between the new version being added and the SDK
builder finishing).

Also add instructions for updating the patches, which now have the
version in the filename.

The instructions are rewritten in the form of shell commands to make
the normal case easier, but some human intervention may still be needed
(e.g. if patches need to be rebased).

BUG=None
TEST=Proofread it

Change-Id: I5480be2172b7d64f2f46aec7d7c27f8023e5b7f7
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1627811
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>