dev-libs/openssl: Uprev to openssl-1.1.1d
Changes vs upstream:
 * Forward-port the local blacklisting patch
 * KEYWORDS="*"
 * cros_use_gcc in setup and configure
 * Hack to keep providing 1.0.2 binaries for transitioning the SDK

Regarding the last point: We'd ideally just replace the 1.0.2 ebuild
with the 1.1.1 one and have rebuilds for dependencies be triggered by
subslot operators. There are complications though: The
host-post-cross packages are installed as binary packages only and
might still link to the old openssl shared objects (e.g. cargo). To
prevent stuff from breaking, this CL adds a new revision of the
openssl-1.0.2 build in the "legacy" slot that gets installed
concurrently with openssl-1.1.1 and just drops the shared objects with
the old SONAME. The idea is to wait until newer host-post-cross
binpkgs are available, at which point we can drop this hack and remove
the "legacy"-slotted openssl-1.0.2 ebuild.

BUG=chromium:649672, chromium:994667
TEST=chroot creation / update is successful

Change-Id: Ia4d8ca4de6b509261fc7167f48718c9019f5314d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1758579
Tested-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>