arc-build: Erase arc-build from DEPEND/RDPEND for arc-mesa
arc-mesa{,-img,-freedreno} depend on arc-build.
But arc-build should not be included in RDEPEND because it don't need to
be included at runtime.
And they don't have to include arc-build in DEPEND either because
arc-build eclass, which is inherited in arc-mesa, depends on arc-build
itself.

BUG=b:130523566
TEST=build for {eve,eve-arcvm,hana,cheza-freedreno}
     mount image, check /opt/google/{vms,containers}/android/

Change-Id: I0d1f67f52c04159d7334372ad33b080a781360f8
Signed-off-by: Kansho Nishida <kansho@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1634131
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>