rust: fix lib install directory
The 1.35.0 version of rust included a change to the bootstrap script
that seems to have moved where the rust compiler shared objects are
installed. The toolchain itself seems to work fine but the proc-macro
unit tests, which are dynamically linked to the compiler shared objects,
where not able to find them in their new location.

This CL fixes this issue by adjusting the toolchain's build
configuration install location for libs such that this toolchain
installs compiler shared objects to the location of previous toolchains:
/usr/lib64.

BUG=chromium:967436
TEST=sdk-tryjob
     run crosvm unit tests

Change-Id: I839edb1e6d8f97b54d3ea3af662e4702eebb730d
Reviewed-on: https://chromium-review.googlesource.com/1633953
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>