cros-workon: remove existing symlink before create
If the symlink to create was already made in the previous run,
symlink_in_place has been creating a symlink in the directory the
existing one points to.
It is not intended, and causes error in case it would overwrite an existing
file.
This change fixes the issue by removing existing symlink (if any) beforehand.

TEST=Run cros_workon_make update_engine twice.
BUG=chromium:967711

Change-Id: Ib74a6e5b5be42c335850e9f41cd0a39cf33942a2
Reviewed-on: https://chromium-review.googlesource.com/1635269
Commit-Ready: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>