glibc: use __builtin_constant_p to skip trivial FORITFY checks
This is only done on functions which are presumed to be hot and
often-used, and which are trivially optimizeable by the compiler.

The hope is that this might both make CrOS faster, and make the cost of
things like automatic initialization less: the compiler can't exactly
reason about arbitrary _chk functions, but it has a much easier time
reasoning about a regular memcpy, for example.

BUG=None
TEST=Tryjobs

Change-Id: Iae794f4e129abe262e7381b419202a490a9de9f7
Reviewed-on: https://chromium-review.googlesource.com/1638320
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: George Burgess <gbiv@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>