chromeos-cr50-scripts: cr50 event log tracking script
The new script is invoked once after startup.

First the script updates the Cr50 flash log time base, so that new
events would be reported with the correct time stamp.

The time stamp of the most recent reported event is saved in
/mnt/stateful_partition/unencrypted/preserve/cr50_flog_timestamp which
makes sure that it survives powerwashes, OS updates, etc. If the file
does not exist, it is created with the value of time stamp of 0.

The script polls the Cr50 to retrieve log entries added after the most
recent reported one. If there are such entries, their timestamp is
saved and the event ID is added to the sparse UMA histogram
Platform.Cr50.FlashLog, and after successful addition the time stamp
is saved as the

BRANCH=none
BUG=b:132287488
TEST=generated a few log entries using Cr50 console command and ran
     the script. Observed the entries showing up in
     chrome://histograms and verified that the timestamps are matching
     real time.

     Verified that there are no duplicate histogram updates when this
     script is re-ran, and that new events show up in the histogram if
     they are added on the Cr50.

     Verified that when running on a DUT with Cr50 version not yet
     supporting flash log, cr50-metrics.conf still completes
     successfully.

Change-Id: Iae47fe015e0fb41f6ff165f4db9f5bdbe53c1bfc
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1649196
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Namyoon Woo <namyoon@chromium.org>