cros-kernel2.eclass: Add devres debugging code for 'devdebug' use flag
This is another knob that can be turned on to debug the devres layer
(all the devm_* calls in the kernel). The overhead is minimal (just some
extra debug printing code), and nothing is actually printed unless you
pass devres.log_devres=1 on the kernel commandline, so this should be
fine to add.

BUG=None
TEST=Build kernel with devdebug use flag, pass devres.log_devres=1 on
commandline and see lots of devres logging

Change-Id: I04bfdeff8e179656e2ff43d185c9763537601cd0
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1650150
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>