eclass/cros-kernel: Add DM_INIT to configs
This is necessary to boot a verity rootfs with the dm-mod.create
commandline. Tack it onto the end of each config so this can be used.

BUG=chromium:960168
TEST=Boot with verity and dm-mod.create="vroot,,0,ro, 0 4710400 verity 0
					8:19 8:19 4096 4096 588800 588800 sha1
					9b0a223aedbf74b06442b0f05fbff33c55edd010
					414b21fba60a1901e23aec373e994942e991d6762631e54a39bc42411f244bd2"
or something similar to this and see the vroot device in 'dmsetup table'

Change-Id: Ide94c54d2dbc2d84d17c6893b39276e5de245342
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1650152
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>