alsa-utils, alsa-lib: mark 1.1.8 as stable
Carry patches over from 1.1.6 (drop any that are already upstream) and
unmask the package. This also adds 2 new patches. One for the Dell dock ucm
and another to remove the veyron-i2s and chtrt5650 ucms as they conflict
and we have no formal solution to deal with that conflict yet.

BUG=b:132848755
TEST=build CRAS and deploy to chromebook

Cq-Depend: chromium:1657029
Change-Id: Id970a2bb922efb45f5bdba5f8caef2035fff79b4
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>