ebuild for libmems and mems_setup
mems_setup is meant to be a boot-time configuration tool.
In its initial incarnation, it also included a bunch of
useful wrapper classes around libiio, because that was
the only logical place they could fit.

The plan, however, is to use those same wrapper classes
to support other clients of libiio in platform2. So, it
makes sense to split mems_setup in two parts at this time:

libmems: a library that wraps libiio and provides useful testing
         mocks and helpers around that, with proper platform2/
         ergonomics, and easily accessible by the rest of the codebase.

mems_setup: a boot-time configuration tool, solely responsible for
            initializating the sensors on a ChromeOS device, and the
            replacement for accelerometer-init.sh.

BUG=chromium:969101
TEST=emerge and test mems_setup after the split

Change-Id: I999a6598147c814a3cd3c73fddd0eac731b9ab34
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1666610
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>