cros-rust: build tests in their own target directory
If the test step was run in a cros-rust ebuild, they would overwrite the
release binary of the previous compile step because they share the same
target directory. This change adds the --target-dir option to the cargo
test command so that a different target directory is selected, fixing
this issue.

TEST=FEATURES=test emerge-nami crosvm
     check that crosvm has features enabled
BUG=None

Change-Id: Idbf1b8addf2bb4c1f14a1ad01eb14ff841a5d307
Reviewed-on: https://chromium-review.googlesource.com/1669437
Tested-by: Zach Reizner <zachr@chromium.org>
Commit-Ready: Zach Reizner <zachr@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>