chromeos-base/crosvm: drop crosvm-usb USE flag
This flag is misleadingly named: it sounds like it should control
whether USB support is enabled, but in fact it actually just controlled
the 'sandboxed-libusb' feature flag, and USB support is always enabled
regardless of the USE flag state.  The USE flag also (incorrectly)
controlled whether the dependency on libusb was added; building without
USE=crosvm-usb would still link against libusb but without declaring the
dependency in the portage dep graph.

Building without sandboxed-libusb enabled won't work at all in the
Chrome OS environment, since we need to be able to pass USB devices via
file descriptor, which is only accepted when crosvm is built with
sandboxed-libusb.  This is reinforced by the fact that crosvm-usb is
enabled by default in the ebuild (via +) and is never disabled in any
existing board profiles.

To avoid confusion and clean up the ebuild, drop the crosvm-usb flag
entirely, and always build with the sandboxed-libsub feature enabled.

BUG=None
TEST=`FEATURES=test emerge-nami chromeos-base/crosvm`

Change-Id: Ie2248c3c4e1ecdef0c9f5a8d5ed38fbfd5b58986
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1682592
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>