sys-boot/coreboot: Provide better diagnostics when config wrong/missing
When the config file template is missing (or wrong), all you get is
a weird message that the build failed to download SeaBIOS. This is
because the default configuration is to build an emulator image with
SeaBIOS as payload.

Detect the "build an emulator" case and throw an error that is more
indicative of where to look.

BUG=none
TEST=emerge-eve coreboot with config.eve removed presents a useful error

Change-Id: I5fafd2e197fd942198c784f16154ad3f48223829
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1683232
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>