exynos5_defconfig: Make it at least boot on linux-5.2
With this defconfig I was able to get peach_pi to at least boot.  I
didn't spend tons of time optimizing / making it work on as many
kernels as possible / looking for missing features, but this could at
least boot a recent mainline which is better than what was there
before.

If someone wants to try to make something better they could see what
else was missing / diff more carefully to the previous version of this
(which will still be in git history).

NOTE: in order to get things booting, I needed to disable hs400, but
that's because I tested on an earlier version of pi that didn't have
the strobe line connected.

BUG=None
TEST=Boot peach_pi on linux-5.2

Change-Id: I02ac6e2ccfbe2c2cfc237f7ea3970e0ea5455d9b
Reviewed-on: https://chromium-review.googlesource.com/1693554
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>