lld: Cherry pick r365759 and r365760 to llvm-next for version script fixes.
This CL cherry-picks two recent LLD fixes on version script
https://reviews.llvm.org/rL365759 and https://reviews.llvm.org/rL365759.

The cherry-pick is using local patches because they are not applied
cleanly to llvm-next because https://reviews.llvm.org/rL365595 changes
all the variable names in LLD to start with a lowercase letter.

The CL is implemented as a epatch_between so if next compiler update picks
up the refactor, we can do normal cherry picks on the patches.

BUG=chromium:982882,chromium:972454
TEST=LLD builds locally. Will test with SDK builder.

Change-Id: Ie4aeeff3182a81a209a9ed701da410d31e3fcea3