binutils: Make a comment about not fixing a known bug.
This patch adds a comment in binutil ebuild: we have fixed a bug
b/134709902 (https://sourceware.org/bugzilla/show_bug.cgi?id=24642)
on upstream AOSP but don't want to apply to Chrome OS because we
are moving away from gold. We leave a comment for future reference.

This is the same as crrev.com/c/1657088. Just put the comments
in 9999 ebuild, so it'll be kept after uprevving.

BUG=b:134709902
TEST=No functional changes.

Change-Id: Ia7271b67f0d5325c38fd87a0a2f5b2180809b333