xkeyboard-config: drop brokenbar from Swiss (German) layout
Historically, the Swiss (German) layout has had two vertical bar
characters:

* Senkrechter Strich (U+007C, vertical line)
* Unterbrochener Strich (U+00A6, broken bar)

The broken bar is a relic from old school mechanical typesetting to
have a character that looked different from uppercase-I and
lowercase-l. Nowadays, the character does not have use for a typical
user and only serves as a point of confusion when shells treat it
differently from the regular vertical bar.

For what it's worth, upstream xkeyboard-config and Mac OS X do not
have this character in their layout. Windows appears to still have the
character (at least in Windows 7).

In the interest of reducing the number of Chrome OS specific patches
to xkeyboard-config and reducing user confusion, we should drop this
patch.

Thanks to ricop@ on eng-misc for helping explain the historical
context of this character.

BUG=chromium:969208
BRANCH=none
TEST=emerged xkeyboard-config, observed patch was not applied

Change-Id: Id8651997d0dc42a138870182913163c793e00690
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1713838
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>