test-dlc: increase the number of test DLC(s)
These changes are required for dlcservice to more thoroughly test
workflows for multiple DLC(s).

The structure of /opt/google/dlc for imageloader.json and table are
exactly the same, only an increase in number of test-dlc(s) from a
single test-dlc to test1-dlc and test2-dlc. It is simple to increase the
number of test{N}-dlc's by changing the local variable n.

The structure of /usr/local/dlc is as follows on DUT:

/usr/local/dlc/
|-- test1-dlc
|   `-- test-package
|       |-- dlcservice_test-dlc.payload
|       `-- dlcservice_test-dlc.payload.json
`-- test2-dlc
    `-- test-package
        |-- dlcservice_test-dlc.payload
        `-- dlcservice_test-dlc.payload.json

Each test DLC will have it's own payload in specific directory along
with its .json payload data for tast tests to leverage.

BUG=none
TEST=emerge-$BOARD test-dlc
TEST=cros deploy $TEST_IP test-dlc # verify /opt/google/dlc + /usr/local/dlc

Cq-Depend:chromium:1749665
Change-Id: Id3a30cd9b4a0d59cd75e15d11b4f7547b7ee566e